requirements.txt
Django==2.2.5
django-mdeditor==0.1.16
Markdown==3.1.1
Pygments==2.4.2
pytz==2019.2
sqlparse==0.3.0
